move various services to the Users folder
