+++ web/app/Services/Users/Authentication.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Wikijump\Services\Authentication;
+namespace Wikijump\Services\Users;
 
 use Wikijump\Models\User;
 use Illuminate\Http\Request;
